+++ CHANGELOG.md
@@ -3,35 +3,59 @@
 ### Breaking changes
 
 * All JSON methods on Realm now only wraps JSONException in RealmException. All other Exceptions are thrown as they are.
+* Removed `HandlerController` from the public API.
+
+### Deprecated
+
+* `RealmConfiguration.setModules()`. Use `RealmConfiguration.modules()` insteasd.
+
+### Enhancements
+
+* `RealmObjectSchema.getPrimaryKey()` (#2636)
+
+### Credits
+
+* Thanks to Brenden Kromhout (@bkromhout) for adding `RealmObjectSchema.getPrimaryKey()`.
+
+## 0.89.1
+
+### Bug fixes
+
+* @PrimaryKey + @Required on String type primary key no longer throws when using copyToRealm or copyToRealmOrUpdate (#2653).
+* Primary key is cleared/changed when calling RealmSchema.remove()/RealmSchema.rename() (#2555).
+* Objects implementing RealmModel can be used as a field of RealmModel/RealmObject (#2654).
 
 ## 0.89.0
 
 ### Breaking changes
 
-* RealmResults.clear() now throws UnsupportedOperationException. Use RealmResults.deleteAllFromRealm() instead.
-* RealmResults.remove(int) now throws UnsupportedOperationException. Use RealmResults.deleteFromRealm() instead.
-* Removed deprecated methods Realm.getTable() from public API.
+* @PrimaryKey field value can now be null for String, Byte, Short, Integer, and Long types. Older Realms should be migrated, using RealmObjectSchema.setNullable(), or by adding the @Required annotation. (#2515).
+* `RealmResults.clear()` now throws UnsupportedOperationException. Use `RealmResults.deleteAllFromRealm()` instead.
+* `RealmResults.remove(int)` now throws UnsupportedOperationException. Use `RealmResults.deleteFromRealm(int)` instead.
+* `RealmResults.sort()` and `RealmList.sort()` now return the sorted result instead of sorting in-place.
+* Removed deprecated method `Realm.getTable()` from public API.
 
 ### Deprecated
 
-* RealmObject.removeFromRealm() in place of RealmObject.deleteFromRealm()
-* Realm.clear(Class) in favour of Realm.delete(Class).
-* DynamicRealm.clear(Class) in place of DynamicRealm.delete(Class).
+* `RealmObject.removeFromRealm()` in place of `RealmObject.deleteFromRealm()`
+* `Realm.clear(Class)` in favour of `Realm.delete(Class)`.
+* `DynamicRealm.clear(Class)` in place of `DynamicRealm.delete(Class)`.
 
 ### Enhancements
 
-* RealmCollection and OrderedRealmCollection have been added. RealmList and RealmResults both implement these interfaces.
-* RealmBaseAdapter now accept an OrderedRealmCollection instead of only RealmResults.
-* RealmObjectSchema.isPrimaryKey(String) (#2440)
-* RealmConfiguration.initialData() can now be used to populate a Realm file before it is used for the first time.
+* Added a `RealmModel` interface that can be used instead of extending `RealmObject`.
+* `RealmCollection` and `OrderedRealmCollection` interfaces have been added. `RealmList` and `RealmResults` both implement these.
+* `RealmBaseAdapter` now accept an `OrderedRealmCollection` instead of only `RealmResults`.
+* `RealmObjectSchema.isPrimaryKey(String)` (#2440)
+* `RealmConfiguration.initialData(Realm.Transaction)` can now be used to populate a Realm file before it is used for the first time.
 
 ### Bug fixes
 
-* RealmObjectSchema.isRequired() and RealmObjectSchema.isNullable() don't throw when the given field name doesn't exist.
+* `RealmObjectSchema.isRequired(String)` and `RealmObjectSchema.isNullable(String)` don't throw when the given field name doesn't exist.
 
 ### Credits
 
-* Thanks to @thesurix for adding RealmConfiguration.initialData().
+* Thanks to @thesurix for adding `RealmConfiguration.initialData()`.
 
 ## 0.88.3
 
@@ -110,6 +134,7 @@
 
 ### Bug fixes
 
+* Error occurring during test and (#2025).
 * Error occurring during test and connectedCheck of unit test example (#1934).
 * Bug in jsonExample (#2092).
 * Multiple calls of RealmResults.distinct() causes to return wrong results (#2198).